refactor(response-example): introduce utility functions for opening collection requests and saving response examples
