+++ tests/response-examples/binary-preview/binary-preview-bru.spec.ts
@@ -1,19 +1,7 @@
-import { test, expect, Page } from '../../../playwright';
+import { test, expect } from '../../../playwright';
 import { buildCommonLocators } from '../../utils/page/locators';
-import { clickResponseAction, expandCollection, expandFolder } from '../../utils/page/actions';
+import { openCollectionRequest, sendAndSaveResposeExample } from '../../utils/page/response-example';
 
-/**
- * Each fixture request POSTs to the local test server's /api/echo/custom
- * endpoint, which echoes back the embedded base64 payload as raw bytes
- * with the requested Content-Type. Saving that response as an example
- * should store the body as binary and render the matching preview.
- *
- * The collection is a temp copy (via collectionFixturePath), so saving
- * examples never mutates the committed fixture.
- *
- * These cases cover the `.bru` collection; the image requests live in the
- * sibling `.yml` collection and are covered by binary-preview-yml.spec.ts.
- */
 const binaryPreviewCases = [
   {
     requestName: 'binary-preview-pdf',
@@ -39,7 +27,6 @@ const binaryPreviewCases = [
     folderName: 'audio',
     exampleName: 'AAC Example',
     previewType: 'audio',
-    // Served as audio/mp4, but the preview trusts the sniffed bytes (M4A signature)
     expectedMime: 'audio/m4a'
   },
   {
@@ -58,45 +45,20 @@ const binaryPreviewCases = [
   }
 ];
 
-const openBinaryPreviewRequest = async (page: Page, folderName: string | undefined, requestName: string) => {
-  const locators = buildCommonLocators(page);
-  await expandCollection(page, 'bru-collection');
-  if (folderName) {
-    await expandFolder(page, folderName);
-    await locators.sidebar.folderRequest(folderName, requestName).click();
-  } else {
-    await locators.sidebar.request(requestName).click();
-  }
-};
-
-const saveResponseAsExample = async (page: Page, requestName: string, exampleName: string) => {
-  const { request, responseExample } = buildCommonLocators(page);
-  await request.sendButton().click();
-  await clickResponseAction(page, 'response-bookmark-btn');
-
-  await responseExample.nameInput().clear();
-  await responseExample.nameInput().fill(exampleName);
-  await page.getByRole('button', { name: 'Create Example' }).click();
-  await expect(responseExample.title()).toHaveText(`${requestName} / ${exampleName}`);
-};
-
-// The app instance (and its temp collection copy) is reused across tests and CI
-// retries in a worker, so example names must be unique per attempt to avoid
-// colliding with an example a failed attempt already created.
 const uniqueExampleName = (exampleName: string, testInfo: { retry: number }) =>
   testInfo.retry ? `${exampleName} (retry ${testInfo.retry})` : exampleName;
 
-test.describe('Binary response example previews', () => {
+test.describe.serial('Binary response example previews', () => {
   for (const { requestName, folderName, exampleName, previewType, expectedMime } of binaryPreviewCases) {
     test(`should preview a saved ${previewType} response (${requestName})`, async ({ pageWithUserData: page }, testInfo) => {
       const savedExampleName = uniqueExampleName(exampleName, testInfo);
 
-      await test.step('Open collection and request', async () => {
-        await openBinaryPreviewRequest(page, folderName, requestName);
+      await test.step('Open collection', async () => {
+        await openCollectionRequest(page, 'bru-collection', folderName, requestName);
       });
 
       await test.step('Send request and save response as example', async () => {
-        await saveResponseAsExample(page, requestName, savedExampleName);
+        await sendAndSaveResposeExample(page, requestName, savedExampleName);
       });
 
       await test.step('Verify the binary preview renders', async () => {
@@ -120,11 +82,11 @@ test.describe('Binary response example previews', () => {
 
   test('should sniff the real content type when the header is mislabeled (binary-preview-mislabeled)', async ({ pageWithUserData: page }, testInfo) => {
     await test.step('Open collection and request', async () => {
-      await openBinaryPreviewRequest(page, undefined, 'binary-preview-mislabeled');
+      await openCollectionRequest(page, 'bru-collection', undefined, 'binary-preview-mislabeled');
     });
 
     await test.step('Send request and save response as example', async () => {
-      await saveResponseAsExample(page, 'binary-preview-mislabeled', uniqueExampleName('Mislabeled Example', testInfo));
+      await sendAndSaveResposeExample(page, 'binary-preview-mislabeled', uniqueExampleName('Mislabeled Example', testInfo));
     });
 
     // The response is PNG bytes served with a text/plain header — the preview
@@ -141,11 +103,11 @@ test.describe('Binary response example previews', () => {
     const locators = buildCommonLocators(page);
 
     await test.step('Open collection and request', async () => {
-      await openBinaryPreviewRequest(page, undefined, 'binary-preview-unknown-binary');
+      await openCollectionRequest(page, 'bru-collection', undefined, 'binary-preview-unknown-binary');
     });
 
     await test.step('Send request and save response as example', async () => {
-      await saveResponseAsExample(page, 'binary-preview-unknown-binary', uniqueExampleName('Unknown Binary Example', testInfo));
+      await sendAndSaveResposeExample(page, 'binary-preview-unknown-binary', uniqueExampleName('Unknown Binary Example', testInfo));
     });
 
     // application/octet-stream bytes with no recognizable signature have no
